+++ python_ipware/__init__.py
@@ -1,2 +1,2 @@
-from .ipware import IpWare  # noqa
+from .python_ipware import IpWare  # noqa
 from .__version__ import __version__  # noqa
