+++ python_ipware/__version__.py
@@ -1,0 +1,1 @@
+__version__ = "2.0.0"
