+++ tests/tests_ipv6.py
@@ -2,7 +2,7 @@ import unittest
 import logging
 from ipaddress import IPv4Address, IPv6Address
 
-from ipware import IpWare
+from python_ipware import IpWare
 
 logging.disable(logging.CRITICAL)  # Disable logging for the entire file
